requirements.txt
python-telegram-bot
requests
pandas
scikit-learn
python-dotenv
